requirements.txt
asgiref==3.8.1
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
lxml==5.4.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pycparser==2.22
pypdfium2==4.30.1
python-docx==1.1.2
sqlparse==0.5.3
typing_extensions==4.13.2
